Schedule update oct 2025

+++ flowback/schedule/tasks.py
@@ -1,4 +1,5 @@
 from celery import shared_task
+from django.utils import timezone
 from django_celery_beat.models import PeriodicTask
 
 from flowback.schedule.models import ScheduleEvent
@@ -11,7 +12,15 @@ def event_notify(event_id: int):
     """
     event = ScheduleEvent.objects.get(id=event_id)
 
-    if not event or not event.repeat_frequency:
+    # Skip notify if scheduled at a later date
+    if event.start_date > timezone.now():
+        return
+
+    # Stop notifying if scheduled after end_date
+    if (not event
+        or (not event.repeat_frequency
+            and event.end_date
+            and event.end_date < timezone.now())):
         PeriodicTask.objects.filter(name=f"schedule_event_{event_id}").delete()
 
     if event.repeat_frequency:
